wip decompress type-2

+++ src/flate.zig
@@ -302,5 +302,5 @@ pub const Flate = struct {
     /// Maximum value for deflate "code length" symbols (non-inclusive)
     pub const cl_symbol_max: usize = 19;
     /// The order to write and read the code length symbol lengths
-    pub const cl_code_order = [_]usize{16, 17, 18, 0, 8, 7, 9, 6, 10, 5, 11, 4, 12, 3, 13, 2, 14, 1, 15};
+    pub const cl_code_order = [_]u16{16, 17, 18, 0, 8, 7, 9, 6, 10, 5, 11, 4, 12, 3, 13, 2, 14, 1, 15};
 };
